extract outstanding reports grouping from the derived value
